Refactor first part
