+++ library/books/book.rb
@@ -1,4 +1,4 @@
-require './rental'
+require './library/rentals/rental'
 
 class Book
   attr_accessor :title, :author, :rentals
