+++ library/people/corrector.rb
