+++ library/people/student.rb
