+++ library/people/teacher.rb
@@ -1,0 +1,12 @@
+require './library/people/person'
+
+class Teacher < Person
+  def initialize(age, specialization, name = 'Unknown')
+    super(age, name)
+    @specialization = specialization
+  end
+
+  def can_use_services?
+    true
+  end
+end
